+++ setupTest/jest.setup.js
@@ -32,3 +32,5 @@ jest.mock('react-native-toast-message', () => ({
 	hide: jest.fn(),
 	setRef: jest.fn(),
 }));
+
+jest.spyOn(console, 'warn').mockImplementation(() => {});