+++ src/index.ts
@@ -3,7 +3,6 @@ import BaseButton from 'atoms/BaseButton';
 import CheckBox from 'atoms/CheckBox';
 import Icon from 'atoms/Icon';
 import Image from 'atoms/Image';
-import Input from 'atoms/Input';
 import List from 'atoms/List';
 import Loading from 'atoms/Loading';
 import RadioButton from 'atoms/RadioButton';
@@ -13,6 +12,7 @@ import SwipeUp from 'atoms/SwipeUp';
 import {SwipeUpFlatList, SwipeUpScrollView, SwipeUpView} from 'atoms/SwipeUp/childComponents';
 import Text from 'atoms/Text';
 import BaseInput from 'atoms/BaseInput';
+import Typography from 'atoms/Typography';
 import Collapsible from 'atoms/Collapsible';
 
 // Molecules
@@ -27,6 +27,7 @@ import {configToast} from 'molecules/Toast/utils';
 import SwipeList from 'molecules/SwipeList';
 import ItemSelectionButton from 'molecules/ItemSelectionButton';
 import MainCardList from 'molecules/MainCardList';
+import Input from 'molecules/Input';
 
 // Organisms
 import LoadingFullScreen from 'organisms/LoadingFullScreen';
@@ -43,10 +44,10 @@ export {
 	CheckBox,
 	Icon,
 	Image,
+	Input,
 	Loading,
 	Svg,
 	StatusChip,
-	Input,
 	palette,
 	LoadingFullScreen,
 	RadioButton,
@@ -70,5 +71,6 @@ export {
 	SwipeItemSelectionList,
 	MainCardList,
 	BaseInput,
+	Typography,
 	Collapsible,
 };